+++ proofs/ssprove/_CoqProject
@@ -44,3 +44,5 @@
 ./handwritten/CoreTheorem.v
 # ./handwritten/ModularTheorem.v
 # ./handwritten/MainTheorem.v
+
+./handwritten/BertieResult.v
